wdbc-kedro+tf/src/requirements.txt
ipython>=7.0.0, <8.0
jupyter>=1.0.0, <2.0
jupyter_client>=5.1.0, <6.0
jupyterlab==0.31.1
kedro==0.16.4
kedro-viz>=3.1.0, <4.0
nbstripout==0.3.3
pytest-cov>=2.5, <3.0
pytest-mock>=1.7.1,<2.0
pytest>=3.4, <4.0
wheel==0.32.2
scikit-learn==0.20.2
tensorflow-cpu
numpy
